common/file.go: Add 64-bit box support
- Fix unit-tests broken by (Box).String() now showing offset as hex.
- Use GetParentBoxName() in all relevant places.
- Add additional parse/seek verbosity.
- Correct which logger is being used in some code that has been moved box<->file.
- Validate that box names are valid.
- Fix a glitch in the TKHD InlineString() output.

+++ common/file_test.go
@@ -78,7 +78,7 @@ func TestFile_readBoxAt_Front(t *testing.T) {
 	boxSize, boxType, err := file.readBoxAt(0)
 	log.PanicIf(err)
 
-	if boxSize != uint32(0x01020304) {
+	if boxSize != int64(0x01020304) {
 		t.Fatalf("Size not correct: (%d)", boxSize)
 	} else if boxType != "abcd" {
 		t.Fatalf("Type not correct: [%s]", boxType)
@@ -100,7 +100,7 @@ func TestFile_readBoxAt_Middle(t *testing.T) {
 	boxSize, boxType, err := file.readBoxAt(4)
 	log.PanicIf(err)
 
-	if boxSize != uint32(0x01020304) {
+	if boxSize != int64(0x01020304) {
 		t.Fatalf("Size not correct: (%d)", boxSize)
 	} else if boxType != "abcd" {
 		t.Fatalf("Type not correct: [%s]", boxType)
